requirements.txt
fastapi
uvicorn
boto3
python-dotenv
pydantic
psycopg2-binary
python-multipart
pypdf
ragas
datasets
opentelemetry-sdk 
opentelemetry-instrumentation-fastapi


